edited Gain::process to take a vector of floats instead of ints

+++ gain.h
@@ -5,5 +5,5 @@
 class Gain
 {
     public:
-    std::vector<int> process(std::vector<int>);
+    std::vector<float> process(std::vector<float>);
 };
